third_party/blink/web_tests/fast/media/mq-color-index-02.html

<!DOCTYPE html>
<html>
<head>
  <title>Media Queries - color-index, invalid numbers</title>
  <link rel="author" title="Rune Lillesveen" href="mailto:rune@opera.com"/>
  <link rel="help" href="http://www.w3.org/TR/css3-mediaqueries/#color-index"/>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
<head>
<body>
  <div id="log"></div>
  <script>

    function test_unknown(query) {
      test(() => {
        assert_false(matchMedia(`${query} or (not ${query})`).matches);
      }, `${query} is unknown`);
      return ;
    }

    test_unknown("(color-index: 0.0)");
    test_unknown("(color-index: -1.0)");
    test_unknown("(color-index: 2.1)");
  </script>
</body>
</html>
